Spliting classes from single files
