+++ src/db/Models/base_model.py
@@ -3,12 +3,10 @@ from datetime import datetime
 from pydantic import BaseModel, ConfigDict, Field
 
 
-class BaseModelDb(BaseModel):
-    model_config = ConfigDict(from_attributes=True)
-
-    last_update: datetime = Field(serialization_alias="lastUpdate", description="Last update of the user")
-    creation: datetime = Field(description="Creation of the user")
-
-
 class NewBaseModelDb(BaseModel):
     model_config = ConfigDict(from_attributes=True)
+
+
+class BaseModelDb(NewBaseModelDb):
+    last_update: datetime = Field(serialization_alias="lastUpdate", description="Last update of the user")
+    creation: datetime = Field(description="Creation of the user")
