+++ src/db/Models/base_query_response_model.py
@@ -1,0 +1,8 @@
+from typing import Sequence
+
+from pydantic import BaseModel, InstanceOf
+
+
+class BaseQueryResponseModel(BaseModel):
+    total_results: int
+    results: Sequence[InstanceOf[BaseModel]]
