+++ src/db/Models/brand_model.py
@@ -1,5 +1,3 @@
-import uuid
-
 from pydantic import BaseModel
