+++ src/db/Models/car_query_response_model.py
@@ -1,0 +1,8 @@
+from typing import Sequence
+
+from db.Models import CarModel
+from db.Models.base_query_response_model import BaseQueryResponseModel
+
+
+class CarQueryResponseModel(BaseQueryResponseModel):
+    results: Sequence[CarModel]
