+++ src/db/Models/register_model.py
@@ -1,0 +1,29 @@
+import uuid
+from datetime import date, datetime, time
+
+from pydantic import BaseModel, Field
+
+from db.Models.base_model import BaseModelDb
+
+
+class RegisterModel(BaseModelDb):
+    id_register: uuid.UUID = Field(default_factory=uuid.uuid4, description="Register id")
+    cd_user: uuid.UUID = Field(description="User id")
+    cd_car: uuid.UUID = Field(description="Car id")
+    distance: float = Field(description="Distance traveled")
+    working_time: time = Field(description="Duration of the trip")
+    mean_consuption: float = Field(description="Mean consuption of the car")
+    number_of_trips: int = Field(description="Number of trips")
+    total_value: float = Field(description="Total value of the trips")
+    register_date: date = Field(description="Date of the trip")
+
+
+class NewRegisterModel(BaseModel):
+    cd_user: uuid.UUID = Field(description="User id")
+    cd_car: uuid.UUID = Field(description="Car id")
+    distance: float = Field(description="Distance traveled")
+    working_time: time = Field(description="Duration of the trip")
+    mean_consuption: float = Field(description="Mean consuption of the car")
+    number_of_trips: int = Field(description="Number of trips")
+    total_value: float = Field(description="Total value of the trips")
+    register_date: date = Field(description="Date of the trip")
